+++ src/Report/StatusCode.php
@@ -32,6 +32,7 @@ enum StatusCode: string
     case AssertionHashedUriMismatch = 'assertion.hashedURI.mismatch';
     case AssertionUndeclared = 'assertion.undeclared';
     case AssertionActionMalformed = 'assertion.action.malformed';
+    case AssertionExternalReferenceMalformed = 'assertion.external-reference.malformed';   // SPEC-032
     case AssertionDataHashMatch = 'assertion.dataHash.match';
     case AssertionDataHashMismatch = 'assertion.dataHash.mismatch';
     case AssertionBmffHashMatch = 'assertion.bmffHash.match';